REF_CODE_REFACTOR: mean and reduce_max refactroing the reference code for mean and reduce_max.
BUG=refactoring existing code.
