+++ tensorflow/lite/micro/kernels/BUILD
@@ -165,6 +165,7 @@ cc_library(
         "quantize.cc",
         "quantize_common.cc",
         "reduce.cc",
+        "reduce_common.cc",
         "reshape.cc",
         "resize_bilinear.cc",
         "resize_nearest_neighbor.cc",
@@ -199,6 +200,7 @@ cc_library(
         "micro_ops.h",
         "pooling.h",
         "quantize.h",
+        "reduce.h",
         "softmax.h",
         "svdf.h",
     ],
