+++ tensorflow/lite/micro/kernels/reduce.h
@@ -1,0 +1,47 @@
+/* Copyright 2021 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+#ifndef TENSORFLOW_LITE_MICRO_KERNELS_REDUCE_H_
+#define TENSORFLOW_LITE_MICRO_KERNELS_REDUCE_H_
+
+#include "tensorflow/lite/c/builtin_op_data.h"
+#include "tensorflow/lite/c/common.h"
+
+namespace tflite {
+
+struct OpDataReduce {
+  int32_t multiplier;
+  int shift;
+  int temp_buffer_idx;
+  int resolved_axis_idx;
+  int input_zp;
+  float input_scale;
+  int output_zp;
+  float output_scale;
+  int num_output_elements;
+};
+
+TfLiteStatus PrepareSimple(TfLiteContext* context, TfLiteNode* node);
+
+TfLiteStatus PrepareMax(TfLiteContext* context, TfLiteNode* node);
+
+TfLiteStatus PrepareMeanOrSum(TfLiteContext* context, TfLiteNode* node);
+
+void ResolveAxis(const int* axis_data, int axis_count,
+                 tflite::MeanParams* op_params);
+
+}  // namespace tflite
+
+#endif // TENSORFLOW_LITE_MICRO_KERNELS_REDUCE_H_
